% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_four}
\alias{theme_four}
\title{Theme four}
\usage{
theme_four()
}
\description{
ggplot theme with an empty dark background and large legend
}
\examples{
test<- data.frame(x=1:10,y=rnorm(10),type=rbinom(10,1,0.3))

ggplot(test,aes(x=x,y=y,colour=type))+
  geom_point()+
  theme_four()
}
