% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_three}
\alias{theme_three}
\title{Theme three}
\usage{
theme_three()
}
\description{
ggplot theme with a clean light colored background
}
\examples{
test<- data.frame(x=1:10,y=rnorm(10),type=rbinom(10,1,0.3))

ggplot(test,aes(x=x,y=y,colour=type))+
  geom_point()+
  theme_three()
}
